fix(react): split WalletLogin to avoid resolving unused chain peers
Previously, WalletLogin.tsx imported both wagmi + @solana/* at module init.
chains="evm" with only EVM peers installed would still crash because the
Solana imports fail to resolve at bundle time.

Splits into a shell (WalletLogin.tsx) + per-chain panels (WalletLogin.EVM,
WalletLogin.Solana). The shell dynamically imports each panel via useEffect
so the Solana bundle is never touched when chains="evm" (and vice versa).

Avoids React.lazy/Suspense so renderToString (used in tests) still works.
Renders a simple fallback per chain until the async import resolves.

Tests updated to exercise (a) the shell layout + fallback rendering and
(b) the panels directly (where signer wiring lives).
